rename starterMods.ts -> welcomeMods.ts
