+++ src/background/welcomeMods.ts
